version 1.0.0
Add versioning and the `-V` flag to report the version.

Add support for system-wide and user-wide initialization files, and the
`-S`, `-U`, and `-P` flags for disabling each initialization file.

Renamed the shell function file and added an installation step to
install the file to `~/.srcrr.inc`. This will make installation of
updates to the functions easier to manage.

Add copyright statements.